GCP cloud ready build and deployments for the current version of the website (#10)
* Minimum changes needed to get docker images running with gcloud run

* A full rework of the development script. This is currently a "build and
run" script with plenty of room for improvement. But I know confidently
that this script serves two operational images+containers for the
website and static content server. So let's get that locked in and move
from there

* some small script changes that prep us for living in google's cute lil cloud

* A microservice for serving the website's static content (#8)

OK. ok. this really needs to be it's own submodule but for now these can remain coupled while i get this project up in the sky. Once it's there
we can revisit a bit of a refactor

* Project formatting (#9)

* dockerfile formatting using dockfmt

* shell formatting with shfmt

* python formatting using black

* some changes to dev-local unrelated to formatting that enable dev to mount containers to local volumes again

* There should be no further issues with deploys of these images

* fixing some things now that static files are hosted properly

* Everything we need for the website to work in the cloud at its current state :3

* simple cloudrun deploy script

* name of this file was bothering me, it'll probably change again